test(ledger): isolate authenticated transport regression
