+++ crates/hyprstream/src/services/ledger/enforcer.rs
@@ -475,7 +475,6 @@ mod tests {
     use hyprstream_crypto::cose_sign::sign_composite;
     use hyprstream_crypto::pq::{ml_dsa_sk_from_seed, ml_dsa_sk_to_vk_bytes, ml_dsa_vk_from_bytes};
     use hyprstream_ledger::{LedgerBackend, MemLedger, Purpose};
-    use parking_lot::Mutex;
 
     fn unit() -> UnitId {
         UnitId {
@@ -860,207 +859,6 @@ mod tests {
             }
             other => panic!("missing authenticated PQ key must be denied, got {other:?}"),
         }
-    }
-
-    /// Exercise the production envelope → JWT → handler path, rather than
-    /// placing a DID-shaped string into a Casbin `Subject` fixture. The
-    /// authenticated ledger holder is the `did:key` derived from the envelope
-    /// signer; the JWT's Casbin user remains the unrelated label `alice`.
-    #[tokio::test]
-    async fn verified_user_transport_did_admits_its_grant_and_refuses_another_signer() {
-        struct AdmissionProbe {
-            enforcer: Arc<LocalEnforcer>,
-            grant_cid: Cid,
-            nonce: u128,
-            authz: SpendAuthorization,
-            signing_key: ed25519_dalek::SigningKey,
-            key_source: Arc<dyn hyprstream_rpc::auth::JwtKeySource>,
-            observed: Arc<Mutex<Vec<(String, Did, Option<DenyReason>)>>>,
-        }
-
-        #[async_trait::async_trait(?Send)]
-        impl hyprstream_rpc::RequestService for AdmissionProbe {
-            async fn handle_request(
-                &self,
-                ctx: &hyprstream_rpc::EnvelopeContext,
-                _payload: &[u8],
-            ) -> anyhow::Result<(Vec<u8>, Option<hyprstream_rpc::Continuation>)> {
-                let subject =
-                    AuthenticatedSubject::from_verified_envelope(ctx).map_err(|error| {
-                        anyhow::anyhow!("ledger identity conversion refused: {error:?}")
-                    })?;
-                let result = self
-                    .enforcer
-                    .admit(
-                        &AdmissionRequest {
-                            authenticated_subject: Some(subject.clone()),
-                            grant_cid: self.grant_cid.clone(),
-                            unit: unit(),
-                            amount: 100,
-                            nonce: self.nonce,
-                            spend_authz: Some(self.authz.clone()),
-                        },
-                        0,
-                    )
-                    .await;
-                let reason = match &result {
-                    AdmissionResult::Admitted { .. } => None,
-                    AdmissionResult::Rejected(rejection) => Some(rejection.reason.clone()),
-                };
-                self.observed
-                    .lock()
-                    .push((ctx.subject().to_string(), subject.did, reason));
-                Ok((Vec::new(), None))
-            }
-
-            fn name(&self) -> &str {
-                "ledger-admission-probe"
-            }
-
-            fn transport(&self) -> &hyprstream_rpc::transport::TransportConfig {
-                static TRANSPORT: std::sync::OnceLock<hyprstream_rpc::transport::TransportConfig> =
-                    std::sync::OnceLock::new();
-                TRANSPORT.get_or_init(|| {
-                    hyprstream_rpc::transport::TransportConfig::inproc("ledger-admission-probe")
-                })
-            }
-
-            fn signing_key(&self) -> ed25519_dalek::SigningKey {
-                self.signing_key.clone()
-            }
-
-            fn jwt_key_source(&self) -> Option<Arc<dyn hyprstream_rpc::auth::JwtKeySource>> {
-                Some(Arc::clone(&self.key_source))
-            }
-
-            fn jwt_verify_policy(&self) -> hyprstream_rpc::crypto::CryptoPolicy {
-                // The regression's JWT is an existing EdDSA CA token. The
-                // envelope and ledger spend authorization are both Hybrid.
-                hyprstream_rpc::crypto::CryptoPolicy::Classical
-            }
-        }
-
-        fn signed_wire(
-            client: &ed25519_dalek::SigningKey,
-            client_pq: &hyprstream_crypto::pq::MlDsaSigningKey,
-            jwt: String,
-            server_kem: &hyprstream_rpc::crypto::hybrid_kem::RecipientPublic,
-        ) -> Vec<u8> {
-            use hyprstream_rpc::ToCapnp as _;
-
-            let response_recipient = hyprstream_rpc::crypto::hybrid_kem::generate_recipient(
-                hyprstream_rpc::crypto::hybrid_kem::SuiteId::HyKemX25519MlKem768,
-            )
-            .expect("response recipient");
-            let request = hyprstream_rpc::envelope::RequestEnvelope::new(Vec::new())
-                .with_jwt_token(jwt)
-                .with_service_domain("ledger-admission-probe")
-                .expect("valid service domain")
-                .with_response_kem_recipient(response_recipient.public());
-            let signed = hyprstream_rpc::envelope::SignedEnvelope::new_signed_encrypted_mesh_kem(
-                request, client, client_pq, server_kem,
-            )
-            .expect("hybrid request envelope");
-            let mut message = capnp::message::Builder::new_default();
-            signed.write_to(
-                &mut message.init_root::<hyprstream_rpc::common_capnp::signed_envelope::Builder>(),
-            );
-            let mut wire = Vec::new();
-            capnp::serialize::write_message(&mut wire, &message).expect("serialize request");
-            wire
-        }
-
-        let (mut enforcer, holder, grant_cid, _debit, _credit, alice_ed, alice_pq) =
-            fixture(1_000).await;
-        enforcer.require_pq = true;
-        let enforcer = Arc::new(enforcer);
-        let nonce = rand_nonce();
-        let (authz, _ed_vk, _pq_vk, _) = signed_authz(
-            &grant_cid,
-            enforcer.cell_identity(),
-            nonce,
-            100,
-            u64::MAX,
-            &alice_ed,
-            &alice_pq,
-        );
-        let bob_ed = ed25519_dalek::SigningKey::from_bytes(&[3u8; 32]);
-        let bob_pq = ml_dsa_sk_from_seed(&[4u8; 32]);
-
-        // This is the same kid-anchored PQ trust shape the dispatcher uses in
-        // production. Both authenticated user keys are bound before any input
-        // is accepted; no request supplies its own PQ verifier.
-        let alice_pq_vk = ml_dsa_vk_from_bytes(&ml_dsa_sk_to_vk_bytes(&alice_pq)).unwrap();
-        let bob_pq_vk = ml_dsa_vk_from_bytes(&ml_dsa_sk_to_vk_bytes(&bob_pq)).unwrap();
-        let mut store = hyprstream_rpc::envelope::KeyedPqTrustStore::new();
-        store.bind(alice_ed.verifying_key().to_bytes(), &alice_pq_vk);
-        store.bind(bob_ed.verifying_key().to_bytes(), &bob_pq_vk);
-        hyprstream_rpc::envelope::install_verify_config(
-            hyprstream_rpc::envelope::EnvelopeVerifyConfig {
-                policy: hyprstream_rpc::crypto::CryptoPolicy::Hybrid,
-                pq_store: Some(Arc::new(store)),
-            },
-        )
-        .expect("install isolated test verification config");
-
-        let ca = ed25519_dalek::SigningKey::from_bytes(&[7u8; 32]);
-        let issuer = "https://ledger-user.test".to_owned();
-        let key_source: Arc<dyn hyprstream_rpc::auth::JwtKeySource> = Arc::new(
-            hyprstream_rpc::auth::ClusterKeySource::new(ca.verifying_key(), issuer.clone()),
-        );
-        let observed = Arc::new(Mutex::new(Vec::new()));
-        let probe = AdmissionProbe {
-            enforcer,
-            grant_cid,
-            nonce,
-            authz,
-            signing_key: ed25519_dalek::SigningKey::from_bytes(&[9u8; 32]),
-            key_source,
-            observed: Arc::clone(&observed),
-        };
-        let server_kem =
-            hyprstream_rpc::node_identity::derive_mesh_kem_recipient(&probe.signing_key)
-                .expect("server KEM recipient");
-        let jwt_for = |user: &str, key: &ed25519_dalek::SigningKey| {
-            let now = chrono::Utc::now().timestamp();
-            hyprstream_rpc::auth::jwt::encode(
-                &hyprstream_rpc::auth::Claims::new(user.to_owned(), now, now + 3_600)
-                    .with_issuer(issuer.clone())
-                    .with_cnf_jwk(&key.verifying_key().to_bytes()),
-                &ca,
-            )
-        };
-
-        for (user, ed, pq) in [("alice", &alice_ed, &alice_pq), ("bob", &bob_ed, &bob_pq)] {
-            hyprstream_rpc::service::dispatch::process_request(
-                &signed_wire(ed, pq, jwt_for(user, ed), &server_kem.public()),
-                &probe,
-                hyprstream_rpc::envelope::EnvelopeVerification::AnySigner,
-                &probe.signing_key,
-                &hyprstream_rpc::envelope::InMemoryNonceCache::new(),
-                hyprstream_rpc::transport::carrier::CarrierContext::iroh(),
-            )
-            .await
-            .expect("verified user request reaches the ledger PEP");
-        }
-
-        let observed = observed.lock();
-        assert_eq!(observed.len(), 2);
-        assert_eq!(
-            observed[0].0, "alice",
-            "JWT/Casbin user still reaches the handler"
-        );
-        assert_eq!(
-            observed[0].1, holder,
-            "holder is the envelope signer's did:key"
-        );
-        assert_eq!(observed[0].2, None, "holder's authenticated grant admits");
-        assert_eq!(observed[1].0, "bob");
-        assert_ne!(
-            observed[1].1, holder,
-            "different signer derives a different DID"
-        );
-        assert_eq!(observed[1].2, Some(DenyReason::HolderMismatch));
     }
 
     #[tokio::test]
